change RPC messages to u16

+++ rpc/core/src/message.rs
@@ -5,12 +5,12 @@ pub struct PingRequest {}
 
 impl Serializer for PingRequest {
     fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        store!(u32, &1, writer)?;
+        store!(u16, &1, writer)?;
         Ok(())
     }
 
     fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
-        let _version = load!(u32, reader)?;
+        let _version = load!(u16, reader)?;
         Ok(Self {})
     }
 }
@@ -20,12 +20,12 @@ pub struct PingResponse {}
 
 impl Serializer for PingResponse {
     fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        store!(u32, &1, writer)?;
+        store!(u16, &1, writer)?;
         Ok(())
     }
 
     fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
-        let _version = load!(u32, reader)?;
+        let _version = load!(u16, reader)?;
         Ok(Self {})
     }
 }
@@ -35,12 +35,12 @@ pub struct GetStatusRequest {}
 
 impl Serializer for GetStatusRequest {
     fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        store!(u32, &1, writer)?;
+        store!(u16, &1, writer)?;
         Ok(())
     }
 
     fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
-        let _version = load!(u32, reader)?;
+        let _version = load!(u16, reader)?;
         Ok(Self {})
     }
 }
@@ -53,14 +53,14 @@ pub struct GetStatusResponse {
 
 impl Serializer for GetStatusResponse {
     fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        store!(u32, &1, writer)?;
+        store!(u16, &1, writer)?;
         store!(String, &self.sparkled_version, writer)?;
         store!(NetworkId, &self.network_id, writer)?;
         Ok(())
     }
 
     fn deserialize<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
-        let _version = load!(u32, reader)?;
+        let _version = load!(u16, reader)?;
         let sparkled_version = load!(String, reader)?;
         let network_id = load!(NetworkId, reader)?;
         Ok(Self {
